@@ -1,5 +1,5 @@
 #The User model, which defines a unique user and all of the properties they have
-class User < ActiveRecord::Base
+class User < ApplicationRecord
   has_many :active_relationships,  class_name:  "Relationship",
                                    foreign_key: "follower_id",
                                    dependent:   :destroy
@@ -50,9 +50,16 @@ class User < ActiveRecord::Base
   validates_attachment :banner, content_type: {content_type: /\Aimage\/.*\Z/}, size: { in: 0..5.megabytes }
 
   after_create :send_signup_email
+  after_validation :clean_paperclip_errors
+
+  def clean_paperclip_errors
+    # Remove avatar/banner file size error if the avatar/banner error's exist
+    errors.delete(:avatar_file_size) unless errors[:avatar].empty?
+    errors.delete(:banner_file_size) unless errors[:banner].empty?
+  end
 
   def send_signup_email
-    UserNotifier.send_signup_email(self).deliver
+    UserNotifier.send_signup_email(self).deliver_now
   end
 
   # Validate the custom_url ...
@@ -74,28 +81,20 @@ class User < ActiveRecord::Base
     !custom_url.empty?
   end
 
-  def to_param
-    has_custom_url? ? custom_url : id.to_s
+  def self.from_param(param)
+    User.find_by!(param.to_s =~ REGEX_USER_ID ? { id: param } : { custom_url: param })
   end
 
   ##########################
   ##### EVENT METHODS ######
   ##########################
 
-  def current_events #return events that are currently going on
-    now = Time.now.in_time_zone("Central Time (US & Canada)")
-    dt_now = DateTime.now
-    busy_events = self.events_in_range(dt_now.beginning_of_day, dt_now.end_of_day) #get events that occur today
-    busy_events = busy_events.select{|event| (event.date <= now and event.end_date >= now)} #get events going on right now
-    return busy_events.sort_by(&:end_date) #return the busy events sorted by which ends soonest
+  def current_events # return events that are currently going on
+    events_in_range(1.day.ago, DateTime.current).select(&:current?).sort_by(&:end_date)
   end
 
   def next_event #returns the next upcoming event within the next day
-    now = Time.now.in_time_zone("Central Time (US & Canada)")
-    dt_now = DateTime.now
-    upcoming_events = self.events_in_range(dt_now.beginning_of_day, dt_now.end_of_day)
-    upcoming_events = upcoming_events.select{|event| (event.date > now and event.date.in_time_zone("Central Time (US & Canada)").to_date == Date.today)} #get future events that occur today
-    return upcoming_events.sort_by(&:date)[0]
+    events_in_range(DateTime.current, 1.day.from_now).min_by(&:date)
   end
 
   def is_busy? #returns whether the user is currently busy (has an event going on)
@@ -104,11 +103,11 @@ class User < ActiveRecord::Base
 
   def events_in_range(start_date_time, end_date_time) #returns all instances of events, including cloned version of repeating events
     #fetch not repeating events first
-    event_instances = events.where(:date => start_date_time...end_date_time, :repeat => nil)
+    event_instances = events.where(:date => start_date_time...end_date_time, :repeat => nil).to_a
 
     #then repeating events
     events.includes(:repeat_exceptions, category: :repeat_exceptions).where.not(repeat: nil).each do |rep_event| #get all repeating events
-      event_instances.concat(rep_event.events_in_range(start_date_time, end_date_time)) #and add them to the event array
+      event_instances.concat(rep_event.events_in_range(start_date_time, end_date_time, home_time_zone)) #and add them to the event array
     end
 
     event_instances = event_instances.sort_by(&:date) #and of course sort by date
@@ -189,11 +188,6 @@ class User < ActiveRecord::Base
     else
       active_relationships.create(followed_id: other_user.id)
     end
-  end
-
-  # TEMP - Forces following, only used for converting from friends to followers
-  def force_follow(other_user)
-    active_relationships.create(followed_id: other_user.id, confirmed: true)
   end
 
   # Unfollows a user.
@@ -313,5 +307,52 @@ class User < ActiveRecord::Base
   ##########################
   ## END GEN USER METHODS ##
   ##########################
+
+  ##########################
+  ##     MISC METHODS     ##
+  ##########################
+
+  # Convert the user into a hash with the least data needed to show search
+  # Recall that clients can see the JSON with a bit of inspection, so only
+  # public information should be included here
+  def convert_to_json
+    user_obj = {} #create a hash representing the user
+
+    # Required fields for search/tokenInput - name and image url
+    user_obj[:name] = self.name
+    user_obj[:image_url] = self.user_avatar(50)
+    user_obj[:id] = self.id
+    user_obj[:model_name] = "User" # specify what type of object this is (used for site search, which handles many object types)
+
+    user_obj #and return the user
+  end
+
+  # Ranks a collection of users for a given search query
+  def self.rank(users, query)
+    query = query.downcase
+
+    users.sort {|a,b| b.rank(query) <=> a.rank(query) } # return users by rank descending (hence sort is flipped)
+  end
+
+  # Ranks the user based on the query
+  # The ranking prioritizes the first name starting with the query
+  # then the middle/last name starting with the query
+  # and finally the query being included at all (handled by SQL)
+  def rank(query)
+    score = 0 # default score is 0
+
+    if name.downcase.starts_with?(query)
+      score = 2
+    elsif name.downcase.include?(" " + query)
+      score = 1
+    end
+
+    return score
+  end
+
+  ##########################
+  ##   END MISC METHODS   ##
+  ##########################
+
 
 end
