@@ -35,6 +35,8 @@ var viewMode = "week"; //either "week" or "month"
 
 var saveEventsTimeout; // timeout for save events so it doesn't happen too often
 
+var PLACEHOLDER_NAME = "<i>Untitled</i>"; // used by newly created categories and events
+
 /****************************/
 /**** DOCUMENT FUNCTIONS ****/
 /****************************/
@@ -47,6 +49,13 @@ $(document).on('page:load', scheduleReady);
 $(document).unbind('page:before-change'); //unbind page before change from last time viewing
 $(document).on('page:before-change', pageChange); //and load again
 
+
+/**
+ * If the user is allowed to leave, returns undefined.
+ * If it is not a good time to leave, returns if the user is sure they want to leave.
+ * run beforeunload
+ * @return {boolean} true if the person wants to leave, false if the user wants to stay, undefined if user can leave
+ */
 function pageChange() //called by Turbolinks before-change
 {
 	if(!isSafeToLeave()) //if the save button is active (they have changes) and this is the user's schedule
@@ -70,7 +79,7 @@ $(window).on('beforeunload', function()
  * by the active state of the Save button on the page; on other pages,
  * it is always safe for the user to leave the page, as there is no data
  * to be saved.
- * @return {Boolean} - returns true if changes are saved, nothing otherwise
+ * @return {boolean} returns true if changes are saved, nothing otherwise
  */
 function isSafeToLeave()
 {
@@ -144,6 +153,12 @@ function ScheduleItem()
 		updatedEvents(this.tempId, "Destroy");
 	};
 
+	/**
+	 * Sets a start date time for the current event
+	 * @param {Date} newStartDateTime - the new start date time that should be set
+	 * @param {boolean} resize - true if the object is being resized, false if the object is being moved
+	 * @param {boolean} userSet - checks if the user is the one who updated the time directly
+	 */
 	this.setStartDateTime = function(newStartDateTime, resize, userSet) //if resize is true, we do not move the end time
 	{
 		if(newStartDateTime.getTime() > this.endDateTime.getTime() && userSet) //if trying to set start before end
@@ -158,6 +173,12 @@ function ScheduleItem()
 			updatedEvents(this.tempId, "setStartDateTime"); // indicate the event was modified, triggering autocomplete
 	};
 
+	/**
+	 * Sets an end date time for the current event
+	 * @param {Date} newEndDateTime - the new end date time that should be set
+	 * @param {boolean} resize - true if the object is being resized, false if the object is being moved
+	 * @param {boolean} userSet - checks if the user is the one who updated the time directly
+	 */
 	this.setEndDateTime = function(newEndDateTime, resize, userSet) //if resize, we don't move the start time
 	{
 		if(newEndDateTime.getTime() < this.startDateTime.getTime() && userSet) //if trying to set end before start
@@ -171,6 +192,10 @@ function ScheduleItem()
 		updatedEvents(this.tempId, "setEndDateTime");
 	};
 
+	/**
+	 * Changes the name of the current event
+	 * @param {String} newName - the new name that should be set
+	 */
 	this.setName = function(newName)
 	{
 		if(this.name != newName) // check for changes
@@ -181,6 +206,19 @@ function ScheduleItem()
 		}
 	};
 
+	/**
+	 * Changes the repeat type of the current event
+	 * @param {String} newRepeatType - new repeat type that should be set
+	 * putting here for the time being
+	 * 1. these will (like the name implies) happen every day/week/month/year
+	 *   `daily`, `weekly`, `monthly`, `yearly`
+	 * 2. events that happen on certain days every week, string will look like this
+	 *   `certain_days-<comma seperated day values>`
+	 *   e.g. certain_days-1,3,5 (1,3,5 is mon wed fri)
+	 * 3. events that are entirely custom are built like so
+	 *   `custom-<amount>-<unit>`
+	 *   e.g. custom-3-weeks (will happen every 3 weeks)
+	 */
 	this.setRepeatType = function(newRepeatType)
 	{
 		if(this.repeatType != newRepeatType) // check for changes
@@ -190,6 +228,10 @@ function ScheduleItem()
 		}
 	};
 
+	/**
+	 * Set the date time where the repeating should start for the current event
+	 * @param {Date} newRepeatStart - where the repeating should start
+	 */
 	this.setRepeatStart = function(newRepeatStart)
 	{
 		if(this.repeatStart != newRepeatStart) // check for changes
@@ -199,6 +241,10 @@ function ScheduleItem()
 		}
 	};
 
+	/**
+	 * Set the date time where the repeating should end for the current event
+	 * @param {Date} newRepeatEnd - where the repeating should end
+	 */
 	this.setRepeatEnd = function(newRepeatEnd)
 	{
 		if(this.repeatEnd != newRepeatEnd) // check for changes
@@ -208,6 +254,10 @@ function ScheduleItem()
 		}
 	};
 
+	/**
+	 * Set the description of the current event
+	 * @param {Date} newDescription - new description
+	 */
 	this.setDescription = function(newDescription)
 	{
 		if(this.description != newDescription) // check for changes
@@ -217,6 +267,10 @@ function ScheduleItem()
 		}
 	}
 
+	/**
+	 * Set the location of the current event
+	 * @param {Date} newLocation - new location
+	 */
 	this.setLocation = function(newLocation)
 	{
 		if(this.location != newLocation) // check for changes
@@ -226,6 +280,11 @@ function ScheduleItem()
 		}
 	}
 
+	/**
+	 * Runs once user has stoped dragging an event, either to resize or move
+	 * @param {JQuery} elem - element that was dragged
+	 * @param {boolean} resize - true if the object is being resized, false if the object is being moved
+	 */
 	this.dragComplete = function(elem, resize)
 	{
 		var dateString = elem.parent().siblings(".col-titler").children(".evnt-fulldate").html();
@@ -242,6 +301,10 @@ function ScheduleItem()
 			updatedEvents(this.tempId, "dragComplete");
 	};
 
+	/**
+	 * Runs once user has stoped resizing an event
+	 * @param {JQuery} elem - element that has been resized
+	 */
 	this.resizeComplete = function(elem)
 	{
 		this.dragComplete(elem, true);
@@ -268,22 +331,21 @@ function ScheduleItem()
 		offsets.push(gridHeight*(this.endDateTime.getMinutes()/60));
 		return offsets;
 	};
-
+	/** Changes height of current event based on the time it takes up */
 	this.updateHeight = function()
 	{
-		this.element().css("height", gridHeight*this.lengthInHours() - border);
-		updatedEvents(this.tempId, "updateHeight");
+		// only update height in view mode's where size indicates duration
+		if(viewMode == "week")
+		{
+			this.element().css("height", gridHeight*this.lengthInHours() - border);
+			updatedEvents(this.tempId, "updateHeight");
+		}
 	};
 
 	/** a way of getting the name that handles untitled */
-	this.getName = function(useHTML)
+	this.getHtmlName = function()
 	{
-		if(this.name)
-			return this.name;
-		else if(useHTML)
-			return "<i>Untitled</i>";
-		else
-			return "Untitled";
+		return this.name ? escapeHtml(this.name) : PLACEHOLDER_NAME;
 	}
 
 	/** Returns the HTML element for this schedule item, or elements if it is repeating */
@@ -301,10 +363,10 @@ function ScheduleItem()
 
 	/**
 	 * Sets the start or end date/time for an event on a user's schedule.
-	 * @param {Boolean}  isStart - Whether or not the Date object being passed in is an event's starting time
+	 * @param {boolean}  isStart - Whether or not the Date object being passed in is an event's starting time
 	 * @param {Date}    dateTime - The date/time this event is being changed to; can be start or end date
 	 * @param {string}   schItem - The jQuery selector for the schedule item being modified
-	 * @param {Boolean}   resize - Whether or not we are resizing the schedule item we're setting the time for
+	 * @param {boolean}   resize - Whether or not we are resizing the schedule item we're setting the time for
 	 */
 	function setDateTime(isStart, dateTime, schItem, resize)
 	{
@@ -341,7 +403,11 @@ function ScheduleItem()
 		}
 
 		elem.attr("time", topDT.getHours() + ":" + paddedMinutes(topDT)); //set the time attribute
-		schItem.tempElement = elem; // update temp element for later populateEvents() calls
+
+		if(viewMode == "month")
+			repopulateEvents();
+		else if(viewMode == "week")
+			schItem.tempElement = elem; // update temp element for later populateEvents() calls - only used by weekly view
 	}
 
 	/**
@@ -349,7 +415,7 @@ function ScheduleItem()
 	 * whether or not to round that result to the nearest number of hours.
 	 * @param  {Date}    start - Starting date for the difference calculation
 	 * @param  {Date}      end - Ending date for the difference calculation
-	 * @param  {Boolean} round - If true, round difference up or down to the nearest hour,
+	 * @param  {boolean} round - If true, round difference up or down to the nearest hour,
 	 *                             rounding up to one if the result of the rounding is zero
 	 * @return {Date}        - the difference between the two given dates, in hours
 	 */
@@ -377,10 +443,15 @@ function ScheduleItem()
 function Category(id)
 {
 	this.id = id; //the id of the category in the db
-	this.name = "<i>Untitled</i>"; //the name of the category, as a string. Defaults to Untitled
+	this.name; //the name of the category, as a string.
 	this.color; //the color of the category, as a CSS acceptable string
 	this.privacy = "private"; //the privacy of the category, right now either private || followers || public
 	this.breaks = []; //an array of the repeat exceptions of this category.
+
+	this.getHtmlName = function()
+	{
+		return this.name ? escapeHtml(this.name) : PLACEHOLDER_NAME;
+	}
 }
 
 /**
@@ -435,11 +506,9 @@ function scheduleReady()
 	if(readOnly) //allow viewing of all events with single click
 	{
 		$(".edit, #repeat, #add-break-event").remove(); //remove repeat functionality, and adding breaks
-		$("#overlay-loc, #overlay-desc, #overlay-title").attr("contenteditable", "false"); //disable editing on location title and description
+		$("#overlay-title").attr("contenteditable", "false"); //disable editing on location title and description
+		$("#overlay-loc, #overlay-desc").prop('disabled', true);
 		$("#time-start, #time-end").attr("readonly", true); //disable editing of time
-		$(".col-snap .sch-evnt").click(function(){
-			editEvent($(this)); //and make it so clicking an event immediately opens it
-		});
 	}
 
 	$("#sch-save").addClass("disabled");
@@ -451,6 +520,11 @@ function scheduleReady()
  */
 function addStartingListeners()
 {
+	// resizes textboxes to give them height based on the content inside
+	$('.auto-resize-vertically').on('input', function () {
+	  textareaSetHeight(this);
+	});
+
 	$(".date-field").datepicker( //show the datepicker when clicking on the field
 	{
 		firstDay: 1, //set Monday as the first day
@@ -566,7 +640,7 @@ function addStartingListeners()
 		}
 	}).click(function()
 	{
-		if($(this).html() == "<i>Untitled</i>")
+		if($(this).html() == PLACEHOLDER_NAME)
 			$(this).text("");
 	})
 	.focusin(function() {
@@ -576,7 +650,7 @@ function addStartingListeners()
 	{
 		// If no name, set to Untitled
 		if($(this).text() == "")
-			$(this).html("<i>Untitled</i>");
+			$(this).html(PLACEHOLDER_NAME);
 
 		removeHighlight();
 	});
@@ -629,8 +703,7 @@ function addStartingListeners()
 	{
 		//TODO: Fix this not working across different days (try noon in your local time)
 
-		//TODO - Fix this reading HTML and parsing it. It's terrible.
-		var dateE = currEvent.element().parent().siblings(".col-titler").children(".evnt-fulldate").html(); //the date the elem is on
+		var dateE = currEvent.startDateTime.toDateString();
 
 		var val = $(this).val();
 
@@ -647,14 +720,13 @@ function addStartingListeners()
 
 		currEvent.setStartDateTime(newDateTime, true, true); //and set!
 		currEvent.updateHeight();
-
 	});
 
 	$("#time-end").change(function()
 	{
 		//TODO: Fix this not working across different days (try noon in your local time)
 
-		var dateE = currEvent.element().parent().siblings(".col-titler").children(".evnt-fulldate").html(); //the date the elem is on
+		var dateE = currEvent.startDateTime.toDateString();
 
 		var val = $(this).val();
 
@@ -668,7 +740,6 @@ function addStartingListeners()
 
 		currEvent.setEndDateTime(newDateTime, true, true);
 		currEvent.updateHeight();
-
 	});
 
 	$("#overlay-desc").focusin(function() {
@@ -676,7 +747,7 @@ function addStartingListeners()
 	})
 	.focusout(function()
 	{
-		currEvent.setDescription($(this).text());
+		currEvent.setDescription($(this).val());
 		removeHighlight();
 	});
 
@@ -685,7 +756,7 @@ function addStartingListeners()
 	})
 	.focusout(function()
 	{
-		currEvent.setLocation($(this).text());
+		currEvent.setLocation($(this).val());
 		removeHighlight();
 	});
 
@@ -700,7 +771,7 @@ function addStartingListeners()
 	})
 	.focusin(function() {
 		// If the name is Untitled, remove it
-		if($(this).html() == "<i>Untitled</i>")
+		if($(this).html() == PLACEHOLDER_NAME)
 			$(this).text("");
 
 		setTimeout(highlightCurrent, 100); // highlight the whole name after focus
@@ -905,12 +976,10 @@ function loadInitialEvents()
 			var time = dateE.getHours() + ":" + paddedMinutes(dateE);
 
 			clone.children(".evnt-title").text(evnt.name);
-			clone.children(".evnt-time.top").text(convertTo12Hour(dateE)).show();
-			clone.children(".evnt-time.bot").text(convertTo12Hour(dateEnd)).show();
 			clone.attr("time", time);
 			clone.attr("event-id", evnt.id);
 			clone.attr("evnt-temp-id", i); //Set the temp id
-			clone.children(".evnt-desc").html(evnt.description);
+			clone.children(".evnt-desc").text(evnt.description);
 
 			scheduleItems[i].tempElement = clone; //Store the element
 
@@ -981,7 +1050,7 @@ function addDrag(selector)
 		scheduleItems[$(this).parent().attr("evnt-temp-id")].setName($(this).text());
 
 		if($(this).text() == "")
-			$(this).html("<i>Untitled</i>");
+			$(this).html(PLACEHOLDER_NAME);
 
 		removeHighlight();
 	});
@@ -1016,7 +1085,7 @@ function addDrag(selector)
 	})
 	.focusin(function() {
 		// If the name is Untitled, remove it
-		if($(this).html() == "<i>Untitled</i>")
+		if($(this).html() == PLACEHOLDER_NAME)
 			$(this).text("");
 
 		setTimeout(highlightCurrent, 100); // highlight the whole name after focus
@@ -1086,6 +1155,10 @@ function addDrag(selector)
 				var currItem = scheduleItems[eventTempId - 1];
 				currItem.startDateTime = new Date($(this).attr("data-date"));
 				currItem.endDateTime = new Date($(this).attr("data-date"));
+
+				// Set end time to 11:59 PM so it's easier to edit the time
+				currItem.endDateTime.setHours(23);
+				currItem.endDateTime.setMinutes(59);
 				repopulateEvents();
 			}
 
@@ -1232,7 +1305,10 @@ function handleClone(elem, ui)
 	addDrag(clone); //and redo dragging
 }
 
-//called on new events dragged from the sidebar
+/**
+ * Called when new events are dragged from the sidebar
+ * @param {jQuery} elem - The element that was dragged
+ */
 function handleNewEvent(elem)
 {
 	var schItem = new ScheduleItem();
@@ -1262,7 +1338,12 @@ function handleNewEvent(elem)
 	eventTempId++;
 }
 
-//Change time while items are being dragged or resized, and also snap to a vertical grid
+/**
+ * Change time while items are being dragged or resized, and also snap to a vertical grid
+ * @param {jQuery} elem - element the time is being updated on
+ * @param {Object} ui - UI object from jQuery drag handler
+ * @param {boolean} resize - true if resizing event, false if moving event
+ */
 function updateTime(elem, ui, resize) //if we're resizing, don't snap, just update time
 {
 	//TODO: Make this really important function not suck
@@ -1337,7 +1418,13 @@ function updateTime(elem, ui, resize) //if we're resizing, don't snap, just upda
 /****************************/
 
 
-//called by next and previous buttons on click
+/**
+ * Moves the calender forward or backward in time
+ * (e.g. by clicking next on weekly view)
+ * @param {Date} newDateObj - new date to start from
+ * @param {boolean} refresh - if true, cleans off all events from scheduler
+ * @param {boolean} startToday - if true starts the weekly view on the current day
+ */
 function addDates(newDateObj, refresh, startToday)
 {
 	refDate = newDateObj; //set the global date to this new
@@ -1369,7 +1456,7 @@ function addDates(newDateObj, refresh, startToday)
 
 		$(".sch-day-col").each(function(index, col)
 		{
-			$(col).attr("data-date", dateToString(currDate));
+			$(col).attr("data-date", verboseDateToString(currDate));
 
 			var fullDate = monthNames[currDate.getMonth()] + " " + currDate.getDate() + ", " + currDate.getFullYear();
 
@@ -1420,7 +1507,7 @@ function addDates(newDateObj, refresh, startToday)
 			if(currDate < todaySimple)
 				tileClass = tileClass + " in-past";
 
-			$("#sch-monthly-view #tiles-cont").append("<div class='" + tileClass + "' data-date='" + dateToString(currDate) + "' >"
+			$("#sch-monthly-view #tiles-cont").append("<div class='" + tileClass + "' data-date='" + verboseDateToString(currDate) + "' >"
 					+ "<div class='inner'>"
 						+ "<div class='day-of-month'>" + currDate.getDate() + "</div>"
 					+ "</div>"
@@ -1439,6 +1526,7 @@ function addDates(newDateObj, refresh, startToday)
 	populateEvents(); // After refreshing the dates, populate the...er...schedule items for this week. As you can see, the terminology still confuses some.
 }
 
+/** Initializes the monthly view calender */
 function initializeMonthlyView()
 {
 	viewMode = "month";
@@ -1451,6 +1539,7 @@ function initializeMonthlyView()
 	addDates(refDate, true);
 }
 
+/** Initializes the weekly view calender */
 function initializeWeeklyView()
 {
 	viewMode = "week";
@@ -1463,13 +1552,23 @@ function initializeWeeklyView()
 	addDates(refDate, true);
 }
 
-//Takes the month number (1 is Jan.) and the year
+/**
+ * Converts a month and a year to a data object
+ * @param {number} month - 1(January) thru 12(December)
+ * @param {number} year - e.g. 2017
+ * @return {Date} date object made from year and month
+ */
 function daysInMonth(month,year)
 {
 	return new Date(year, month, 0).getDate();
 }
 
-//returns the date the schedule starts on as well as whether it's in this month
+/**
+ * Gets the date the schedule starts on
+ * @param {Date} dateObj - user inputed date used as a reference point to where the schedule should start
+ * @param {boolean} useMonth - if true, sets date to first day of month
+ * @return {Object} object consisting of the start date of an event, and if its start was in the last month
+ */
 function getStartDate(dateObj, useMonth)
 {
 	var copyDate = cloneDate(dateObj);
@@ -1499,33 +1598,44 @@ function getStartDate(dateObj, useMonth)
 	return {startDate: copyDate, lastMonth: lastMonth}
 }
 
-/**
- * Clears events from the schedule before running populateEvents(). Used when the schedule gets updated
- */
+/** Clears events from the schedule before running populateEvents(). Used when the schedule gets updated */
 function repopulateEvents()
 {
 	$("#sch-holder .sch-evnt, #sch-holder .sch-month-evnt").remove(); // remove week and month events
 	populateEvents(); // and then populate events
 }
 
-//Populate the events in the current week from the hashmap
+/** Fills in events in the current week or month, loads from the scheduleItems hash */
 function populateEvents()
 {
-	function place(eventObject, i)
+	/**
+	 * Place an event on the calender
+	 * @param {ScheduleItem} eventObject - The Schedule event object to be placed on the schedule
+	 * @param {number} visibleDate - the index of the date the event falls on in the currently visible dates
+	 */
+	function place(eventObject, visibleDate)
 	{
 		var color = categories[eventObject.categoryId].color;
 		var currentElem = eventObject.tempElement.clone();
 
 		if(viewMode == "week")
 		{
+			// Setup the UI element's color, text, and height to represent the schedule item
 			currentElem.css("background-color", color);
-			currentElem.find(".evnt-title").html(eventObject.getName(true));
+			currentElem.find(".evnt-title").html(eventObject.getHtmlName());
+			currentElem.find(".evnt-time.top").text(convertTo12Hour(eventObject.startDateTime));
+			currentElem.find(".evnt-time.bot").text(convertTo12Hour(eventObject.endDateTime));
+			currentElem.css("height", gridHeight*eventObject.lengthInHours() - border);
+			currentElem.css("top", eventObject.getTop());
+			currentElem.attr("evnt-temp-id", eventObject.tempId);
+
+			// Add the event
 			$(".sch-day-col:eq(" + i + ") .col-snap").append(currentElem);
 		}
 		else if(viewMode == "month")
 		{
 			var className = "";
-			if(eventObject.name == "<i>Private</i>")
+			if(eventObject.name == "Private")
 				className = " private";
 
 			var eventId = "";
@@ -1539,7 +1649,7 @@ function populateEvents()
 			$(".sch-day-tile:eq(" + i + ") .inner").append("<div class='sch-month-evnt" + className + "' evnt-temp-id='" + eventObject.tempId
 				+ "' " + eventId + " data-id='" + eventObject.categoryId + "' data-hour='" + eventObject.startDateTime.getHours() + "' style='color: "
 				+  color +  "; color: " + color + ";'>"
-					+ "<span class='evnt-title'>" + eventObject.getName(true) + "</span>"
+					+ "<span class='evnt-title'>" + eventObject.getHtmlName() + "</span>"
 					+ "<div class='time'>"
 						+ datesToTimeRange(eventObject.startDateTime, eventObject.endDateTime)
 					+ "</div>"
@@ -1683,6 +1793,7 @@ function populateEvents()
 			monthlyTileDroppable();
 		}
 
+		/** Make each monthly event dragable (e.g. sidebar)  */
 		function monthlyEventDraggable()
 		{
 			$(".sch-month-evnt").draggable(
@@ -1726,6 +1837,7 @@ function populateEvents()
 			});
 		}
 
+		/** Make each monthly event dropable into the schedule (e.g. moving sidebar event into schedule)  */
 		function monthlyTileDroppable() {
 			$(".sch-day-tile").droppable(
 			{
@@ -1755,7 +1867,11 @@ function populateEvents()
 	}
 }
 
-//Edit an event's title inline (without the overlay)
+/**
+ * Edit an event's title inline (without the overlay)
+ * @param {ScheduleItem} event - item to edit title of
+ * @param {JQuery} elem - element to edit title of
+ */
 function editEventTitle(event, elem)
 {
 	//return if this is in the sidebar
@@ -1771,7 +1887,11 @@ function editEventTitle(event, elem)
 	$(elem).siblings(".sch-evnt-save").css("display","inline");
 }
 
-//Edit a category using the category overlay
+
+/**
+ * Edit a category using the category overlay
+ * @param {JQuery} elem - category element to update
+ */
 function editCategory(elem)
 {
 	var categoryId = $(elem).attr("data-id");
@@ -1797,7 +1917,7 @@ function editCategory(elem)
 	else //if the color was null or empty remove the background-color
 		$(".cat-top-overlay").css("background-color",""); */
 
-	$(".cat-overlay-title").html(currCategory.name);
+	$(".cat-overlay-title").html(currCategory.getHtmlName());
 	$("#cat-overlay-box").attr("data-id", categoryId);
 
 	$(".color-swatch").removeClass("selected");
@@ -1809,7 +1929,10 @@ function editCategory(elem)
 	});
 }
 
-//show the event editing overlay
+/**
+ * Edit an event using the event overlay
+ * @param {JQuery} elem - event element to update
+ */
 function editEvent(elem)
 {
 	var editingEvent = $(document.activeElement).hasClass("evnt-title");
@@ -1854,13 +1977,13 @@ function editEvent(elem)
 
 		UIManager.slideInShowOverlay("#event-overlay-box");
 
-		$("#overlay-title").html(currEvent.name || "<i>Untitled</i>");
+		$("#overlay-title").html(currEvent.getHtmlName());
 		$("#overlay-color-bar").css("background-color", categories[currEvent.categoryId].color);
 
 		var desc = currEvent.description || ""; //in case the description is null
 		var loc = currEvent.location || ""; //in case the location is null
-		$("#overlay-desc").html(desc);
-		$("#overlay-loc").html(loc);
+		$("#overlay-desc").val(desc);
+		$("#overlay-loc").val(loc);
 
 		if(desc.length == 0 && readOnly) //if this is readOnly and there is no description
 			$("#overlay-desc, #desc-title").hide(); //hide the field and the title
@@ -1878,10 +2001,17 @@ function editEvent(elem)
 		//$(".overlay-time").html(convertTo12Hour(time.split(":")) + " - " + convertTo12Hour(endTime.split(":")));
 		$("#time-start").val(convertTo12HourFromArray(startArr));
 		$("#time-end").val(convertTo12HourFromArray(endArr));
+
+		// resize the textareas to the appropriate size
+		$('.auto-resize-vertically').each(function () {
+			//check if the textbox contains a new line or else it takes up 2 unnessisary lines
+		  textareaSetHeight(this);
+			$(this).css('overflow-y', 'hidden');
+		});
 	}
 }
 
-//Show the overlay for creating a new break
+/** Show the overlay for creating a new break */
 function showBreakCreateOverlay()
 {
 	$("#break-overlay-box input").val(""); //clear all inputs
@@ -1892,9 +2022,12 @@ function setupInvitedUsers() {
 	/* TODO retrieve users from associated event_invite obj */
 }
 
-// Sets up an overlay to add breaks. If managing,
-// it is actually for editing and deleting breaks rather
-// than enabling or disabling breaks on the currente vent
+/**
+ * Sets up an overlay to add breaks. If managing,
+ * it is actually for editing and deleting breaks rather
+ * than enabling or disabling breaks on the currente event
+ * @param {boolean} managing - if true, show 'managing breaks', if false show 'adding breaks'
+ */
 function setupBreakAddOverlay(managing)
 {
 	var currObj;
@@ -1929,7 +2062,8 @@ function setupBreakAddOverlay(managing)
 		}
 
 
-		$("#break-cont").append("<div class='break-elem " +  classAdd +"' data-id='" + id + "' >"
+		// Prepend each break so that the most recently created break created comes first in the list
+		$("#break-cont").prepend("<div class='break-elem " +  classAdd +"' data-id='" + id + "' >"
 				+ checkbox
 				+ breakInstance.name + " | " + dateToString(breakInstance.startDate) + " | " + dateToString(breakInstance.endDate)
 			+ "</div>");
@@ -1944,6 +2078,7 @@ function setupBreakAddOverlay(managing)
 	if(!managing)
 		setupBreakClickHandlers();
 
+	/** Sets up toggles for if the event or category has breaks enabled */
 	function setupBreakClickHandlers()
 	{
 		$(".break-elem").click(function()
@@ -1981,13 +2116,21 @@ function setupBreakAddOverlay(managing)
 	}
 }
 
-//Update the color of the category overlay from a color being picked
+/**
+ * Update the color of the category overlay from a color being picked
+ * @param {JQuery} elem - element to take background color from
+ */
 function changeCategoryColor(elem)
 {
 	$(".cat-top-overlay").css("background-color",$(elem).css("background-color"));
 }
 
-//Setup properties of a place schedule item from the db, setting position and height
+/**
+ * Setup properties of a place schedule item from the db, setting position and height
+ * @param {JQuery} elem -
+ * @param {number} hours - the height of the object is proportional to the hours
+ * @param {number} lengthHours - the amount of hours an element takes up
+ */
 function placeInSchedule(elem, hours, lengthHours)
 {
 	//console.log("Length: " + lengthHours);
@@ -1995,7 +2138,11 @@ function placeInSchedule(elem, hours, lengthHours)
 	$(elem).css("top", hours); //set the top position by gridHeight times the hour
 }
 
-// Events were updated. Called by any modification of an event, which triggers auto saving
+/**
+ * Events were updated. Called by any modification of an event, which triggers auto saving
+ * @param {number} eventId - id of the event being modified
+ * @param {msg} string - message to show when events were updated... not currently used
+ */
 function updatedEvents(eventId, msg)
 {
 	// console.log("Events were updated!" + msg);
@@ -2014,6 +2161,7 @@ function updatedEvents(eventId, msg)
 /*** JSON SERVER METHODS ****/
 /****************************/
 
+/** Saves all events */
 function saveEvents()
 {
 	if($("#sch-save").hasClass("disabled") || $("#sch-save").hasClass("loading")) //if the save button is disabled or already saving
@@ -2067,6 +2215,11 @@ function saveEvents()
 	});
 }
 
+/**
+ * Deletes an event
+ * @param {Event} event - event from preforming an action... stops propagation of this event
+ * @param {JQuery} elem -element to delete
+ */
 function deleteEvent(event, elem)
 {
 	event.stopImmediatePropagation();
@@ -2096,7 +2249,7 @@ function deleteEvent(event, elem)
 		confirmUI("Are you sure you want to delete this event?", deleteEventProper);
 	}
 
-	// Deletes a single event among a repeating set by making a new repeat break and applying it
+	/** Deletes a single event among a repeating set by making a new repeat break and applying it */
 	function deleteSingleEvent()
 	{
 		var tempId = $(elem).parent().attr("evnt-temp-id");
@@ -2130,20 +2283,29 @@ function deleteEvent(event, elem)
 			createBreak(breakTitle, breakDateString, breakDateString, function(newBreak)
 			{
 				newBreakId = newBreak.id;
+				applyNewBreak(); // apply new break after server responds with our break ID
 			});
 		}
+		else
+		{
+			applyNewBreak();
+		}
 
-		// Now that we have a break that will make this event be skipped, add it and update UI accordingly
-		schItem.breaks.push(newBreakId);
+		// Add the new break to the schedule item and repopulate the schedule to apply it
+		function applyNewBreak()
+		{
+			// Now that we have a break that will make this event be skipped, add it and update UI accordingly
+			schItem.breaks.push(newBreakId);
 
-		updatedEvents(schItem.tempId, "breaks"); //mark that the events changed to enable saving
+			updatedEvents(schItem.tempId, "breaks"); //mark that the events changed to enable saving
 
-		repopulateEvents();
+			repopulateEvents();
 
-		UIManager.slideOutHideOverlay("#evnt-delete.overlay-box");
+			UIManager.slideOutHideOverlay("#evnt-delete.overlay-box");
+		}
 	}
 
-	// Deletes the associated event object, like the old delete. This gets rid of all items repeating
+	/** Deletes the associated event object, like the old delete. This gets rid of all items repeating */
 	function deleteEventProper()
 	{
 		var eId = $(elem).parent().attr("event-id");
@@ -2177,12 +2339,13 @@ function deleteEvent(event, elem)
 	}
 }
 
+/** Creates category */
 function createCategory()
 {
 	$.ajax({
 		url: "/create_category",
 		type: "POST",
-		data: {name: "<i>Untitled</i>", user_id: userId, group_id: groupID, color: "silver"},
+		data: {name: "", user_id: userId, group_id: groupID, color: "silver"},
 		success: function(resp)
 		{
 			console.log("Create category complete.");
@@ -2192,7 +2355,7 @@ function createCategory()
 			newCat.show();
 			newCat.attr("data-id", resp.id);
 			newCat.attr("privacy", "private");
-			newCat.find(".evnt-title").html(resp.name);
+			newCat.find(".evnt-title").html(resp.name || PLACEHOLDER_NAME);
 			newCat.attr("id", "");
 			addDrag();
 			// TODO - Make saving the sideHTML a function, as this line is called so many times
@@ -2213,6 +2376,12 @@ function createCategory()
 	});
 }
 
+/**
+ * Deletes category
+ * @param {Event} event - event from action
+ * @param {Jquery} elem - element of category to be deleted
+ * @param {number} id - id of data in element to remove from database
+ */
 function deleteCategory(event, elem, id)
 {
 	confirmUI("Are you sure you want to delete this category?", function(confirmed)
@@ -2247,19 +2416,28 @@ function deleteCategory(event, elem, id)
 	});
 }
 
+/**
+ * Saves Category
+ * @param {Event} event - event from action
+ * @param {Jquery} elem - element of category to be saved
+ * @param {number} id - id of data in element to add to database
+ */
 function saveCategory(event,elem,id)
 {
+	// uses dom to determine if the category has been given an actual name.
+	var categoryName = ( $(".cat-overlay-title").html() === PLACEHOLDER_NAME ? "" : $(".cat-overlay-title").text());
+
 	$.ajax({
 		url: "/create_category",
 		type: "POST",
-		data: {name: $(".cat-overlay-title").html(), id: id, color: $(".cat-top-overlay").css("background-color"),
+		data: {name: categoryName, id: id, color: $(".cat-top-overlay").css("background-color"),
 			privacy: currCategory.privacy, breaks: currCategory.breaks, group_id: groupID},
 		success: function(resp)
 		{
 			console.log("Update category complete.");
 
 			// TODO - Literally what is this doing? These should be functions
-			currCategory.name = $(".cat-overlay-title").html();
+			currCategory.name = $(".cat-overlay-title").text();
 			$("#sch-sidebar .sch-evnt[data-id=" + id + "]").find(".evnt-title").html($(".cat-overlay-title").html()); //Update name in sidebar
 			$(".sch-evnt[data-id=" + id + "]").css("background-color", $(".cat-top-overlay").css("background-color")); //Update color of events
 			sideHTML = $("#sch-tiles").html(); //the sidebar html for restoration upon drops
@@ -2274,10 +2452,13 @@ function saveCategory(event,elem,id)
 	});
 }
 
-// Creates a new break with a given name and starting at startDate
-// and ending at endDate, where both are strings.
-// Fires the passed in callback when the break has been made, passing
-// the break as a parameter
+/**
+ * Creates a new break
+ * @param {String} name - name of break
+ * @param {Date} startDate - start date for the break
+ * @param {Date} endDate - end date for the break
+ * @param {function} callback - once the date has been created run this function
+ */
 function createBreak(name, startDate, endDate, callback)
 {
 	// console.log("Make the break: " + name + ", " + startDate + ", " + endDate);
@@ -2327,13 +2508,22 @@ function createBreak(name, startDate, endDate, callback)
 /***** HELPER METHODS *******/
 /****************************/
 
-//converts a date from 24 hour to 12 hour time string format
+/**
+ * Converts a date from 24 hour to 12 hour time string format
+ * @param {Date} date - date to convert to 12 hour time format
+ * @return {String} date in 12 hour time format
+ */
 function convertTo12Hour(date)
 {
 	var timeArr = [date.getHours(), paddedMinutes(date)]; //and reset the field
 	return convertTo12HourFromArray(timeArr);
 }
 
+/**
+ * Converts a date from time array to 12 hour time string format
+ * @param {Array} timeArr - array to convert to 12 hour time format
+ * @return {String} date in 12 hour time format
+ */
 function convertTo12HourFromArray(timeArr)
 {
 	if(timeArr[0] >= 12)
@@ -2342,7 +2532,7 @@ function convertTo12HourFromArray(timeArr)
 			timeArr[0] -= 12;
 
 		if(timeArr[0] == 0)
-			timeArr[0] == "00";
+			timeArr[0] = "00";
 
 		return timeArr.join(":") + " PM";
 	}
@@ -2352,13 +2542,17 @@ function convertTo12HourFromArray(timeArr)
 			timeArr[0] = 12;
 
 		if(timeArr[0] == 0)
-			timeArr[0] == "00";
+			timeArr[0] = "00";
 
 		return timeArr.join(":") + " AM";
 	}
 }
 
-//returns whether the element is in a schedule column (basically has it been placed in the schedule)
+/**
+ * Checks whether the element is in a schedule column (basically has it been placed in the schedule)
+ * @param {JQuery} elem - element to check
+ * @return {boolean} true if element is in schedule, false if not
+ */
 function inColumn(elem)
 {
 	var class_data = elem.parent().attr("class"); //get the parent's class data
@@ -2368,54 +2562,98 @@ function inColumn(elem)
 		return false;
 }
 
-function setHeight(getElem, setElem, hoursLength) //get the height of getElem and set the height of setElem if the height is not a proper height (divisible by gridheight)
+/**
+ * Set the height of an element if the height of another element
+ * is not a proper height (divisible by gridheight)
+ * @param {JQuery} getElem - element to check the height of
+ * @param {JQuery} setElem - element to set height of
+ * @param {number} hoursLength - hours to set the height too
+ */
+function setHeight(getElem, setElem, hoursLength)
 {
 	var height = parseFloat($(getElem).css("height"));
 	if((height+border)%gridHeight != 0)
 		$(setElem).css("height", (gridHeight*hoursLength)-border);
 }
 
-//returns the minutes of a date in padded form (e.g. 03 instead of just 3)
+/**
+ * Returns the minutes of a date
+ * @param {Date} date - date to get minutes from
+ * @return {String} minutes in padded form (e.g. 03 instead of just 3)
+ */
 function paddedMinutes(date)
 {
 	var minutes = (date.getMinutes() < 10? '0' : '') + date.getMinutes(); //add zero the the beginning of minutes if less than 10
 	return minutes;
 }
 
-//zero pads a number to two digits (9 -> 09, 1 -> 01, 13 -> 13) used for zero padded dates and times (e.g. 2:04 pm type things)
+/**
+ * Zero pads a number to two digits
+ * @param {number} num - number to be zero padded
+ * @return {String} zero padded number (e.g. 3 to 03 or 13 to 13)
+ */
 function paddedNumber(num)
 {
 	var paddedNum = (num < 10? '0' : '') + num; //add zero the the beginning of minutes if less than 10
 	return paddedNum;
 }
 
-//removes cursor highlight on page
+/** Removes cursor highlight on page */
 function removeHighlight()
 {
 	window.getSelection().removeAllRanges();
 }
 
-//highlight the entirety of the field currently selected (that the user has cursor in)
+/** Highlight the entirety of the field currently selected (that the user has cursor in).
+ * Runs HTMLInputElement.select if an input is in focus, otherwise runs 'selectAll' on document.
+ */
 function highlightCurrent()
 {
-	document.execCommand('selectAll',false,null);
+	if($("textarea:focus").length > 0 || $("input:focus").length > 0)
+		$(":focus").select();
+	else
+		document.execCommand('selectAll',false,null);
 }
 
-//creates a clone of the date
+/**
+ * Creates a clone of the date
+ * @param {Date} date - date to clone
+ * @return {Date} clone of date
+ */
 function cloneDate(date)
 {
 	return new Date(date.getTime());
 }
 
-// converts a date string from dashes to slashes (e.g. 2016-10-25 to 2016/10/25)
-// This is needed as browsers don't like dash date formats much, but it's how Ruby prints dates by default
-// On Chrome, dashes with dates are interpreted as the ISO format, and are used in UTC, while Firefox just refuses the date at all
+/**
+ * Updates a textarea element's height based on the content passed to display in it.
+ * Used specifically for event editing on My Schedule, this will update the size of the
+ * description or location text area height dependent on content.
+ * @param {Jquery} elem - The textarea element in question.
+ */
+function textareaSetHeight(elem)
+{
+	$(elem).attr('rows', 1);
+	$(elem).height('auto').height(elem.scrollHeight);
+}
+
+/**
+ * Converts a date string from dashes to slashes (e.g. 2016-10-25 to 2016/10/25).
+ * This is needed as browsers don't like dash date formats much, but it's how Ruby prints dates by default.
+ * On Chrome, dashes with dates are interpreted as the ISO format, and are used in UTC, while Firefox just refuses the date at all.
+ * @param {String} dateString - date with slashes
+ * @return {String} date without slashes
+ */
 function dateFromDashesToSlashes(dateString)
 {
 	return dateString.split("-").join("/");
 }
 
-//convert a date into a standard string format, with no zero padding in M/D/YY format (e.g. 6/2/16)
+/**
+ * Convert a date into a string without zero padding
+ * @param {Date} date - date to be converted to string
+ * @return {String} date in the standard string format, with no zero padding in M/D/YY format (e.g. 6/2/16)
+ */
 function dateToString(date)
 {
 	if(!date || !(date instanceof Date)) //if the date is null or not a date object
@@ -2430,8 +2668,11 @@ function dateToString(date)
 	return dateString; //and return
 }
 
-
-//Converts a date oject to a date string in the format of MM/DD/YYYY, always printing zero padding if needed (e.g. 06/02/2016)
+/**
+ * Convert a date into a string with zero padding
+ * @param {Date} date - date to be converted to string
+ * @return {String} date string in the format of MM/DD/YYYY, always printing zero padding if needed (e.g. 06/02/2016)
+ */
 function verboseDateToString(date)
 {
 	if(!date || !(date instanceof Date)) //if the date is null or not a date object
@@ -2447,12 +2688,22 @@ function verboseDateToString(date)
    return (monthStr[1]?monthStr:"0"+monthStr[0]) + "/" + (dateStr[1]?dateStr:"0"+dateStr[0]) + "/" + yearStr; // padding
 }
 
-
+/**
+ * Converts a date from 24 hour to 12 hour time string format
+ * @param {Date} date - date to convert to 12 hour time format
+ * @return {String} date in 12 hour time format
+ */
 function dateToTimeString(date)
 {
 	return convertTo12Hour(date);
 }
 
+/**
+ * Takes to dates and makes a string to express the range between them
+ * @param {Date} startDate - start date
+ * @param {Date} endDate - end date
+ * @return {String} date range (e.g. 12:00AM to 3:00PM)
+ */
 function datesToTimeRange(startDate, endDate)
 {
 	return dateToTimeString(startDate) + " to " + dateToTimeString(endDate);
@@ -2466,10 +2717,13 @@ function datesToTimeRange(startDate, endDate)
 /**** HTML TIED METHODS *****/
 /****************************/
 
-//Used by the next and previous buttons to change the part of the schedule being shown
-//If forward is true, the schedules moves forward one week, otherwise back one week
-//Worth noting that when the schedule loads, the first day is the current day, not the Monday of that week
-//so that case is accounted for to move the schedule forward to the next Monday
+/**
+ * Used by the next and previous buttons to change the part of the schedule being shown
+ * If forward is true, the schedules moves forward one week, otherwise back one week
+ * Worth noting that when the schedule loads, the first day is the current day, not the Monday of that week
+ * so that case is accounted for to move the schedule forward to the next Monday
+ * @param {boolean} forward - true if the forward button was pressed
+ */
 function moveWeek(forward)
 {
 	var newDate;
