@@ -1,40 +1,7 @@
+require 'tasky'
+
 class PagesController < ApplicationController
-  before_filter  :authorize_admin, :only => [:promote, :admin, :sandbox, :destroy_user, :admin_user_info] #authorize admin on all of these pages
-
-  def promote #promote or demote users admin status
-    @user = User.find(params[:id])
-    if(current_user and current_user.admin)
-      if params[:de] == "true" #if demoting
-        @user.admin = false
-      else
-        @user.admin = true
-      end
-      @user.save
-    end
-
-    #redirect_to "/users"
-    render :json => {"action" => "promote", "uid" => params[:id]}
-  end
-
-  def destroy_user
-    user = User.find(params[:id])
-
-    if current_user.admin
-      user.destroy
-    end
-
-    redirect_to "/admin"
-  end
-
-  def admin_user_info
-    @user = User.find(params[:id])
-
-    @account_creation_time = @user.created_at.strftime('%Y-%m-%d at %H:%M')
-    @last_update_time = @user.updated_at.strftime('%Y-%m-%d at %H:%M')
-    @last_sign_in_time = @user.last_sign_in_at.strftime('%Y-%m-%d at %H:%M')
-    @most_recent_sign_in_time = @user.current_sign_in_at.strftime('%Y-%m-%d at %H:%M')
-    @user_groups = UsersGroup.where(user_id: @user.id)
-  end
+  before_action :authorize_admin!, only: [:admin, :sandbox]
 
   def admin #admin page
     @now = Time.zone.now
@@ -43,5 +10,30 @@ class PagesController < ApplicationController
     @past_month_users = User.where('created_at >= ?', Time.zone.now - 1.months).group(:created_at).count
     @past_month_events = Event.where('created_at >= ?', Time.zone.now - 1.months).group(:created_at).count
     @past_month_events_modified = Event.where('created_at >= ?', Time.zone.now - 1.months).group(:updated_at).count
+  end
+
+  # Runs predefined server commands requested from the admin panel.
+  def run_command
+    cmd = case params[:button_id]
+          when "run-jsdoc"               then "npm run jsdoc --silent"
+          when "run-js-unit-tests"       then "npm run teaspoon --silent"
+          when "run-js-acceptance-tests" then "npm run acceptance-tests --silent"
+          when "run-rails-unit-tests"    then "npm run minitest --silent"
+          end
+
+    task_id = Tasky::run cmd
+    render json: params.merge(task_id: task_id)
+  rescue Tasky::CommandError => e
+    render json: params.merge(cmd_error: e.inspect)
+  end
+
+  def check_if_command_is_finished
+    task = Tasky::fetch_task params[:task_id]
+
+    if task.finished?
+      render json: {log: (task.success? ? "SUCCESS" : task.error_log)}
+    else
+      render json: {check_again: true}
+    end
   end
 end
