@@ -9,39 +9,43 @@ require 'test_helper'
 class PagesControllerTest < ActionController::TestCase
   include Devise::Test::ControllerHelpers
 
-  #Admin Page Security Tests
+  def setup
+    @viktor, @norm, @putin = users(:viktor, :norm, :putin)
+  end
+
+  # Admin Page Security Tests
   test "should not get admin if not signed in" do
     get :admin
     assert_response :redirect
   end
 
   test "should not be able to go to admin if user is not admin" do
-  	sign_in users(:norm)
-  	get :admin
-  	assert_response :redirect
+    sign_in @norm
+    get :admin
+    assert_response :redirect
   end
 
   test "should be able to go to admin if user is admin" do
-  	sign_in users(:viktor)
-  	get :admin
-  	assert_response :success
+    sign_in @viktor
+    get :admin
+    assert_response :success
   end
 
-  #Sandbox Page Security Tests
+  # Sandbox Page Security Tests
   test "should not get sandbox if not signed in" do
-  	get :sandbox
-  	assert_response :redirect
+    get :sandbox
+    assert_response :redirect
   end
 
   test "should not be able to go to sandbox if user is not admin" do
-  	sign_in users(:norm)
-  	get :sandbox
-  	assert_response :redirect
+    sign_in @norm
+    get :sandbox
+    assert_response :redirect
   end
 
   test "should be able to go to sandbox if user is admin" do
-  	sign_in users(:viktor)
-  	get :sandbox
-  	assert_response :success
+    sign_in @viktor
+    get :sandbox
+    assert_response :success
   end
 end
